+++ extras/tests/MixedConfiguration/string_length_size_4.cpp
@@ -26,6 +26,16 @@ TEST_CASE("ARDUINOJSON_STRING_LENGTH_SIZE == 4") {
     }
   }
 
+  SECTION("set(MsgPackExtension)") {
+    SECTION("returns true if size >= 65532") {
+      auto str = std::string(65532, '?');
+      auto result = doc.set(MsgPackExtension(1, str.data(), str.size()));
+
+      REQUIRE(result == true);
+      REQUIRE(doc.overflowed() == false);
+    }
+  }
+
   SECTION("deserializeJson()") {
     SECTION("returns Ok if string length >= 65536") {
       auto input = "\"" + std::string(65536, '?') + "\"";
@@ -47,6 +57,14 @@ TEST_CASE("ARDUINOJSON_STRING_LENGTH_SIZE == 4") {
 
     SECTION("returns Ok if binary size >= 65536") {
       auto input = "\xc5\xff\xff" + std::string(65536, '?');
+
+      auto err = deserializeMsgPack(doc, input);
+
+      REQUIRE(err == DeserializationError::Ok);
+    }
+
+    SECTION("returns Ok if extension size >= 65532") {
+      auto input = "\xc8\xff\xfb\x01" + std::string(65532, '?');
 
       auto err = deserializeMsgPack(doc, input);
 
@@ -78,5 +96,32 @@ TEST_CASE("ARDUINOJSON_STRING_LENGTH_SIZE == 4") {
 
     REQUIRE(result == 5 + str.size());
     REQUIRE(output == std::string("\xc6\x00\x01\x00\x00", 5) + str);
+  }
+
+  SECTION("ext 32 deserialization") {
+    auto str = std::string(65536, '?');
+    auto input = std::string("\xc9\x00\x01\x00\x00\x2a", 6) + str;
+
+    auto err = deserializeMsgPack(doc, input);
+
+    REQUIRE(err == DeserializationError::Ok);
+    REQUIRE(doc.is<MsgPackExtension>());
+    auto value = doc.as<MsgPackExtension>();
+    REQUIRE(value.type() == 42);
+    REQUIRE(value.size() == 65536);
+    REQUIRE(value.data() != nullptr);
+    REQUIRE(std::string(reinterpret_cast<const char*>(value.data()),
+                        value.size()) == str);
+  }
+
+  SECTION("ext 32 serialization") {
+    auto str = std::string(65536, '?');
+    doc.set(MsgPackExtension(42, str.data(), str.size()));
+
+    std::string output;
+    auto result = serializeMsgPack(doc, output);
+
+    REQUIRE(result == 6 + str.size());
+    REQUIRE(output == std::string("\xc9\x00\x01\x00\x00\x2a", 6) + str);
   }
 }